refactor: headerwithunitselector and spacing
