+++ apps/gateway-ui/src/components/index.tsx
@@ -1,6 +1,5 @@
 export { GatewayCard } from './GatewayCard';
 export { WithdrawCard } from './WithdrawCard';
-export { FederationCard } from './FederationCard';
 export { ConnectFederationModal } from './ConnectFederationModal';
 export { InfoCard } from './InfoCard';
 export { DepositCard } from './DepositCard';
